App in Running Stage , Solved a Bug
