+++ assets/data/database.json
@@ -1,1 +1,1 @@
-[{"data": {"title": "tfhjjyjtitit722222222", "time": 0.05, "alarm": true}, "time": "2022-06-08 07:20:05.954148", "state": "Completed", "progress": 0}, {"data": {"title": "fthfhf", "time": 0.05, "alarm": true}, "time": "2022-06-08 07:23:05.144100", "state": "Completed", "progress": 0}, {"data": {"title": "fhfhfjhfj", "time": 0.05, "alarm": true}, "time": "2022-06-08 07:23:08.632642", "state": "Completed", "progress": 0}, {"data": {"title": "fjhyfjyjjfyj", "time": 0.05, "alarm": true}, "time": "2022-06-08 07:23:11.782596", "state": "Completed", "progress": 0}]
+[{"data": {"title": "fthfhf", "time": 0.05, "alarm": true}, "time": "2022-06-08 07:23:05.144100", "state": "Completed", "progress": 0}]
